requirements.txt
dearpygui==2.0.0
glfw==2.8.0
numpy==2.2.1
Pillow==11.1.0
PyOpenGL==3.1.7
pyserial==3.5
typing_extensions==4.12.2
yostlabs==2025.1.9
